[libc++][test] Makes pthread a flag.
Using pthread may affect the compilation too not only the linking.
Therefore pthread is changed from a link flag to a generic flag used for
both compilation and linking.

This fixes an issue when using modules since the use of the pthread flag
makes the BMI files not compatible.

Reviewed By: #libc, ldionne

Differential Revision: https://reviews.llvm.org/D149453

NOKEYCHECK=True
GitOrigin-RevId: d58ba7a4d145c60039d9a87f7265e56716e3bedd