[libcxx] [test] Make set_windows_crt_report_mode.h more explicit
This header is included when building with a debug CRT in
MSVC/Clang-cl environments. By default, failed asserts with the
debug CRT pops up a blocking dialog box alerting the user about
the failed assert. When running more than one test in an automated
fashion, this isn't ideal.

This header tries to run initializers to set the behaviour of the
failed asserts to print a message to the console, just like the
default is in release mode.

This is previously done by setting the reporting mode to
_CRTDBG_MODE_DEBUG, which means outputting to the debugger's
output window. In some setups, this is enough for making it work,
but in others it instead can pop up a dialog asking for which
debugger to use.

Instead set the mode explicitly to _CRTDBG_MODE_FILE and set the
destination to be explicitly to stderr.

For setups where the previous code worked correctly, it doesn't make
any difference other than that a failed assert prints an additional
"abort() has been called" message that wasn't printed before.

Differential Revision: https://reviews.llvm.org/D155823

NOKEYCHECK=True
GitOrigin-RevId: cd1b8be8de91bc1c43bac3eea7ebf3b5643b031c