[libc++] Remove broken self test for the libc++ Lit format
The empty.sh.cpp test never tested what it was intended to test, because
it did contain an unexpected RUN: command. This was discovered in
https://reviews.llvm.org/D154987 while trying to land an unrelated change.
Since there is no reliable way to test what I was trying to test from
the libc++ test suite, just remove the test.

NOKEYCHECK=True
GitOrigin-RevId: 64558826ed060eeeb22b92585c0c2a2a3461606c