[libc++][CI] Moves modules build to C++23.
The std module can be tested in the generic C++23 build. This removes the
special module build.

Note it is possible to enable modules automatically in CMake, but that
would fail in the "parts disabled" builds; they have not been properly
been converted to modules yet.

Depends on D157625

Reviewed By: #libc, ldionne

Differential Revision: https://reviews.llvm.org/D157744

NOKEYCHECK=True
GitOrigin-RevId: c1d49ce67a3ff3c0678ae3e39d846312c9a5633a